Add get product by id endpoint

+++ OnlineStore.Application/Dtos/ProductImageDto.cs
@@ -3,6 +3,6 @@
     public sealed record ProductImageDto
     {
         public string Url { get; set; } = null!;
-        public int ImageOrder { get; set; }
+        public short ImageOrder { get; set; }
     }
 }
